# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Social Network for Music: Backend",
    "description": "The backend for SNM project",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/requireInfo/{kind}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/logout": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  }
}